closes #16: Source location is now reported properly.
For backends that support source location such as class, file, line number, the source of the caller of the slf4s API is now reported instead of the source within the slf4s implementation.

TODO: This change depends on SLF4J API 1.6.x, so probably the SLF4S minor version should be incremented.